Prep for day2 addition / update day1:problem2
Originally in (move), I was returning a vector with (filterv) to be
"firendlier" in how my non-functional brain sees data. In all reality
though, using (filter)'s return of a lazy sequence was all fine as well.
I was overthinking how to work with it since the functions to grab
what's inside a lazy sequence are somewhat limited. But, since I only
needed the first occurance from the filter, (first) works on such a
collection so I was able to grab the vecotr inside of it with no issue.
Furthermore, there's so many ways to skin cats in Clojure and I was
nesting (when)s when a (and) seemed more appropriate. I have some
blogging to do as I learn this stuff ;)